Eva - Syntax formatting has been modularized. It now uses exchangeable SyntaxRegexes objects to determine how to apply coloring to a string.
